Fix sign/unsigned compiler warning

+++ src/serlio/modifiers/PRTModifierAction.cpp
@@ -662,7 +662,7 @@ void PRTModifierAction::updateDynamicEnums() {
 					const size_t cutoffIndex = currStringView.find_first_of(L"\r\n");
 					currStringView = currStringView.substr(0, cutoffIndex);
 
-					const MString mCurrString(currStringView.data(), currStringView.length());
+					const MString mCurrString(currStringView.data(), static_cast<int>(currStringView.length()));
 					MCHECK(e.mAttr.addField(mCurrString, enumIndex));
 				}
 				break;
